Added DNS server with TCP support

+++ src/stat.c
@@ -56,7 +56,7 @@ void stat_request_end(stat_t *s, size_t resp_len, ev_tstamp latency)
     s->responses_size += resp_len;
     s->responses++;
     // NOLINTNEXTLINE(bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions)
-    s->query_times_sum += (latency * 1000);
+    s->query_times_sum += (uint64_t)(latency * 1000);
   }
 }
